backend/requirements.txt
fastapi==0.115.0
uvicorn==0.23.1
redis==5.0.8
feedparser==6.0.11
pydantic==1.10.17
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.43
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.9
google-generativeai==0.7.2
timescale-vector==0.0.7
websockets==13.1

